api/requirements.txt
Django==1.11.7
Pillow==4.2.1
PyJWT==1.5.3
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
django-braces==1.12.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
future==0.16.0
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
olefile==0.44
psycopg2==2.7.3.2
pycryptodomex==3.4.7
pyjwkest==1.4.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
rest-social-auth==1.2.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.5.0
sorl-thumbnail==12.3
urllib3==1.22
